# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "This is vehicle spare parts dealers API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tokens/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new User",
                "description": "User registration",
                "operationId": "userRegister",
                "requestBody": {
                    "description": "Request body for User Registration",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "example@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "example123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "example123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/tokens/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "User login",
                "description": "User login Authentication",
                "operationId": "userLogin",
                "requestBody": {
                    "description": "Request body for User login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "example@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "example123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/api/tokens/delete/{id}'": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete User",
                "description": "Delete User",
                "operationId": "userDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/tokens/update/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user by Id",
                "description": "Update a user record by Id",
                "operationId": "userUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of user to Update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User Updated Successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/api/category/store": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a product category",
                "description": "Category creation",
                "operationId": "categoryStore",
                "requestBody": {
                    "description": "Request body for creating product category",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/category/delete/{id}'": {
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete Category by Id",
                "description": "Delete Category",
                "operationId": "categoryDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Category deleted successfully"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
    }
}